Move files to src, add unit test to check if logged messaged as correct
